added arguments to myListsFragment
