+++ app/src/main/res/drawable/bg_item_book.xml
@@ -1,4 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <selector xmlns:android="http://schemas.android.com/apk/res/android">
-
+    <item android:drawable="@color/grey" android:state_selected="false" />
+    <item android:drawable="@color/green" android:state_selected="true" />
 </selector>
