+++ app/src/main/res/layout/item_list_my_books.xml
@@ -2,10 +2,11 @@
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/colorCL"
     android:layout_width="match_parent"
     android:layout_height="100dp"
-    android:orientation="horizontal"
-    android:id="@+id/colorCL">
+    android:background="@drawable/bg_item_book"
+    android:layout_marginBottom="8dp">
 
 
     <ImageView
@@ -25,20 +26,21 @@
         android:layout_marginStart="40dp"
         android:layout_marginTop="16dp"
         android:layout_toRightOf="@+id/bookCover"
-        android:text="NAME OF BOOK"
         android:textSize="22sp"
+        app:layout_constraintBottom_toTopOf="@id/authorName"
         app:layout_constraintStart_toEndOf="@+id/bookCover"
         app:layout_constraintTop_toTopOf="parent"
-        tools:layout_editor_absoluteY="4dp" />
+        tools:layout_editor_absoluteY="4dp"
+        tools:text="NAME OF BOOK" />
 
     <TextView
         android:id="@+id/authorName"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginStart="40dp"
-        android:layout_marginTop="8dp"
-        android:text="NAME OF AUTHOR"
         app:layout_constraintStart_toEndOf="@+id/bookCover"
-        app:layout_constraintTop_toBottomOf="@+id/bookName" />
+        app:layout_constraintTop_toBottomOf="@+id/bookName"
+        app:layout_constraintVertical_chainStyle="packed"
+        tools:text="NAME OF AUTHOR" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
